feat(web): startViewer library entry with pluggable report sources
